requirements.txt
streamlit
langchain-google-genai
langchain-community
chromadb
pypdf
python-dotenv
langchain-text-splitters
langchain-classic
langchain-core
flashrank
pandas